Move navigation functions to helpers package
