+++ src/components/ScrollOffsetContextProvider.tsx
@@ -2,9 +2,9 @@ import type {ParamListBase} from '@react-navigation/native';
 import React, {createContext, useCallback, useEffect, useMemo, useRef} from 'react';
 import {withOnyx} from 'react-native-onyx';
 import usePrevious from '@hooks/usePrevious';
+import {isSidebarScreenName} from '@libs/Navigation/helpers';
 import type {PlatformStackRouteProp} from '@libs/Navigation/PlatformStackNavigation/types';
 import type {NavigationPartialRoute, State} from '@libs/Navigation/types';
-import {isSidebarScreenName} from '@libs/NavigationUtils';
 import ONYXKEYS from '@src/ONYXKEYS';
 import SCREENS from '@src/SCREENS';
 import type {PriorityMode} from '@src/types/onyx';
