+++ src/libs/Navigation/AppNavigator/FreezeWrapper/getIsScreenBlurred/index.ts
@@ -1,5 +1,5 @@
 import type {NavigationState} from '@react-navigation/native';
-import {isFullScreenName} from '@libs/NavigationUtils';
+import {isFullScreenName} from '@libs/Navigation/helpers';
 
 function getIsScreenBlurred(state: NavigationState, currentRouteKey: string) {
     const lastFullScreenRoute = state.routes.findLast((route) => isFullScreenName(route.name));
