+++ src/libs/Navigation/helpers/index.ts
@@ -1,14 +1,14 @@
 export * from './createNormalizedConfigs';
-export {default as customGetPathFromState} from './customGetPathFromState';
-export {default as getOnboardingAdaptedState} from './getOnboardingAdaptedState';
-export {default as getParamsFromRoute} from './getParamsFromRoute';
-export {default as normalizePath} from './normalizePath';
-export {default as replacePathInNestedState} from './replacePathInNestedState';
+export * from './isNavigatorName';
+export * from './linkTo/types';
+export {default as closeRHPFlow} from './closeRHPFlow';
 export {default as createNormalizedConfigs} from './createNormalizedConfigs';
 export {default as createSplitNavigator} from './createSplitNavigator';
-export {default as getAdaptedStateFromPath} from './getAdaptedStateFromPath';
+export {default as customGetPathFromState} from './customGetPathFromState';
 export {default as extractPolicyIDFromQuery} from './extractPolicyIDFromQuery';
-export {default as closeRHPFlow} from './closeRHPFlow';
+export {default as getAdaptedStateFromPath} from './getAdaptedStateFromPath';
+export {default as getOnboardingAdaptedState} from './getOnboardingAdaptedState';
+export {default as getParamsFromRoute} from './getParamsFromRoute';
 export {default as getPolicyIDFromState} from './getPolicyIDFromState';
 export {default as getStateFromPath} from './getStateFromPath';
 export {default as getTopmostReportParam} from './getTopmostReportParam';
@@ -16,6 +16,11 @@ export {default as getTopmostRouteName} from './getTopmostRouteName';
 export {default as isReportOpenInRHP} from './isReportOpenInRHP';
 export {default as isSearchTopmostFullScreenRoute} from './isSearchTopmostFullScreenRoute';
 export {default as isSideModalNavigator} from './isSideModalNavigator';
+export {default as linkTo} from './linkTo';
+export {default as getMinimalAction} from './linkTo/getMinimalAction';
+export {default as normalizePath} from './normalizePath';
+export {default as replacePathInNestedState} from './replacePathInNestedState';
 export {default as setNavigationActionToMicrotaskQueue} from './setNavigationActionToMicrotaskQueue';
+export {default as setupCustomAndroidBackHandler} from './setupCustomAndroidBackHandler';
 export {default as shouldOpenOnAdminRoom} from './shouldOpenOnAdminRoom';
 export {default as shouldPreventDeeplinkPrompt} from './shouldPreventDeeplinkPrompt';
