+++ src/libs/Navigation/helpers/linkTo/getMinimalAction.ts
