+++ src/libs/Navigation/helpers/linkTo/index.ts
@@ -1,8 +1,9 @@
 import {getActionFromState} from '@react-navigation/core';
 import type {NavigationContainerRef, NavigationState, PartialState, StackActionType} from '@react-navigation/native';
 import {findFocusedRoute, StackActions} from '@react-navigation/native';
-import {getStateFromPath, normalizePath} from '@libs/Navigation/helpers';
 import {getMatchingFullScreenRoute, isFullScreenName} from '@libs/Navigation/helpers/getAdaptedStateFromPath';
+import getStateFromPath from '@libs/Navigation/helpers/getStateFromPath';
+import normalizePath from '@libs/Navigation/helpers/normalizePath';
 import {shallowCompare} from '@libs/ObjectUtils';
 import {extractPolicyIDFromPath, getPathWithoutPolicyID} from '@libs/PolicyUtils';
 import linkingConfig from '@navigation/linkingConfig';
