+++ src/libs/ReportUtils.ts
@@ -74,9 +74,9 @@ import * as Localize from './Localize';
 import Log from './Log';
 import {isEmailPublicDomain} from './LoginUtils';
 import ModifiedExpenseMessage from './ModifiedExpenseMessage';
+import {isFullScreenName} from './Navigation/helpers/isNavigatorName';
 import linkingConfig from './Navigation/linkingConfig';
 import Navigation, {navigationRef} from './Navigation/Navigation';
-import {isFullScreenName} from './NavigationUtils';
 import * as NumberUtils from './NumberUtils';
 import Parser from './Parser';
 import Permissions from './Permissions';
