+++ src/pages/AddPersonalBankAccountPage.tsx
@@ -10,8 +10,8 @@ import ScreenWrapper from '@components/ScreenWrapper';
 import useLocalize from '@hooks/useLocalize';
 import useThemeStyles from '@hooks/useThemeStyles';
 import getPlaidOAuthReceivedRedirectURI from '@libs/getPlaidOAuthReceivedRedirectURI';
+import {isFullScreenName} from '@libs/Navigation/helpers';
 import Navigation, {navigationRef} from '@libs/Navigation/Navigation';
-import {isFullScreenName} from '@libs/NavigationUtils';
 import * as BankAccounts from '@userActions/BankAccounts';
 import * as PaymentMethods from '@userActions/PaymentMethods';
 import NAVIGATORS from '@src/NAVIGATORS';
